Remove redundant setting of x29 to 0.
This instruction was a useless leftover from when I didn't know fp was x29. x29
is used to hold the dynamic vector in this code, so setting it to 0 at the
outset is useless. It is set to 0 before branching to _start_c at the end there,
and that is simpler.