Fix memmove logic bug in remaining architectures.
I don't always copy code, but when I do, I ensure it contains a critical error.
In this case, I just used the wrong branch to go branch to memcpy, and I did
that in all the architectures. Aarch64 additionally had the error that it used a
signed comparison instead of an unsigned one. So that is fixed everywhere now.