Add wcwidth().
I researched how this works in glibc and musl, and decided to go with the glibc
ruleset, if only because that one is understandable. It works by special-casing
two characters, then making rules about them based on their Unicode properties.
I had to download two more Unicode data files, and I decided to just collect the
characters based on type, then hand-edit the files I got in this way for the
special cases.

The function itself is defined by POSIX to return -1 for non-printable
characters, but 0 for the null character. That's why I special case them. In the
end, this is just a question of sets: There is a set of characters for which -1
is returned, a set for which 0 is returned, a set for which 2 is returned, and
for everything else, 1 is returned.

POSIX defines that the application shall ensure the parameter is actually a wide
character, which I presume to mean that iswcntrl(wc) || iswprint(wc) is true
(printing and control characters partition the character space). And the set of
characters for which -1 is returned is exactly equal to the set of control
characters except for the null character, so there is no separate set for that.

musl has some interesting partitioning of the higher planes here, but my
representation of sets can efficiently represent large ranges, so this is
unnecessary. It also special cases all code points ending in FFFE, but those
aren't Unicode characters, so I don't care.