Reject non-portable rounds specs in SHA crypt.
Rich Felker had claimed that the reference implementation allows for
non-portable rounds specifications, which I had rejected previously. But
thinking about it more, if a negative rounds number is specified, the
interpretation depends on the value of ULONG_MAX, and that would indeed create a
specification that leads to different rounds values for different machines.

This is entirely impossible if the rounds spec starts with a single digit. So I
now require that, and reject it otherwise.