Add missing generic wstring API.
Most of these are pretty obvious. Because I only support 4-byte wchar_t, there
are no clever algorithms for most of these. wcsstr() for example is just the
two-way algorithm, with no special cases for short needles. wcsspn() and
wcscspn() are just the quadratic-time algo, because a set representation on
4-byte integers would be very impractical. And so on.