Add file orientation and basic wchar FILE API.
Each file has an orientation and an encoding rule, as per POSIX. The orientation
is the mode value for fwide(), and has almost no further bearing on the library.
The C standard basically says that all the unspecified cases left over from this
implementation are undefined behavior.

The orientation is automatically set in __stdio_read() and __stdio_write(). For
__stdio_read(), this means it is set on first buffer fill, but for write it only
happens on first buffer flush. Oh well. The expression used changes the
orientation from 0 to -1, keeps it at -1 if it already is there, but leaves a
value of 1 unchanged.

All the basic wide character I/O functions call fwide(f, 1) before doing
anything, and refuse to work if that doesn't return that the orientation is now
wide.

fwide() for its part will then set the file's locale depending on the current
LC_CTYPE setting (looking at MB_CUR_MAX). POSIX calls this "encoding rule", and
basically specifies this behavior. I will simply set the thread-local locale to
the file's locale for the duration of the call.

Wide-oriented streams basically define iterators over a string of wide
characters. Simultaneously, C defines that their associated files contain
multi-byte strings. So the only question remaining is whether their buffers are
wide characters or multibyte characters. I have decided to just use multi-byte
characters, i.e. the translation happens in fgetwc() and fputwc(). This allows
me to isolate the translation stuff in a small number of functions, and not have
to put it in the write backends. Though it makes more advanced functions later
harder.