Add *wprintf() API.
Once more, all variants eventually converge on vfwprintf(), just like I did for
the multi-byte version. However, not to repeat all the code, the numeric
conversions all consist only of ASCII characters, and so no difference exists
for them being printed directly into the file, or being converted to wide
characters first and printed via fputwc(). So all of that can just be done via
fprintf(), as far as I'm concerned.

The *swprintf() variants are a bit weird. They require me to error out if the
space is insufficient, which means a caller can only retry with a larger buffer
and hope it's correct now. Also, because the buffer space is defined to be
multi-byte characters, they have to be reconverted to wide characters. So for
example the literal parts of a format string will be converted wide->mb in
vfwprintf() first, then mb->wide in string_write(). But the alternative would be
the weird macro construction I see in other libcs, so I just keep the slight
wastefulness.