Add reallocarray().
This really just calls realloc(), but it causes an error if the multiplication
overflows. I saw that GCC will generate better code if I remove the zero check
for b, but I cannot do that, because I would be leaving in undefined behavior.
It is bike shedding anyway.

The thorny issue of realloc(p, 0) is left to realloc() itself to deal with. I
think I shall have to revisit the topic sooner or later.