scandir: Add closedir() calls and EOVERFLOW for multiplication.
While buflen is increasing in single steps, capacity is not, and can
overflow. Luckily, unsigned overflow is defined, so I can just check if
the number got smaller (after overflow on addition, the result will be
smaller than either input, so in particular it will be smaller than the
capacity on input).

And closedir() is really necessary to cleanup stuff.