Add SysV IPC functions.
ftok() is taken from the other libcs, they all do the same thing. For the
syscalls, I had a choice to either always try the modern syscall and fall back
to SYS_ipc where available, or else always do SYS_ipc. Since the latter is the
older syscall and not available everywhere, the cost of calling it exclusively
is probably smaller than the cost of the fallback. So I always call SYS_ipc
where available.