Disable ABI warnings on GCC.
I have some internal functions that return a union containing a long double. The
ABI for passing such a thing on x86_64 changed with GCC 4.4. Of course, those
functions are not part of the ABI surface, and have no way to be, but GCC would
rather annoy me with those useless warnings than hold its peace. So I'm going
the other way, shutting up the offending warning.