Add better ilogb alg and overflow detection.
Trying to allocate more than half an addres space is undefined behavior anyway,
but the code never tried to avoid zero-sized allocations. So now I explicitly
detect the overflow condition. I could fail with ENOMEM, since almost surely not
enough memory would be available, but I want to make sure the application
understands that this is not a temporary condition, nor one that can be remedied
by buying more RAM.

Also, it is not necessary to use the bit-smear technique when I have a_clz
available.