Improve handling of errno in glob().
A recent thread on the musl mailing list made me look at my own handling of
errno. readdir() is another one of those APIs that seemed like a good idea at
the time, isn't it? You call it until it returns a NULL pointer, and if you want
to know if that return value is due to EOF or due to error, then you set errno
to 0 before the call and check it afterwards.

Except you need to call readdir() multiple times, and errno is a global variable
that gets changed at the drop of a hat. Which also means that closedir() might
just set it. It doesn't in this implementation, but I want to isolate different
components as best as I can, and nothing in glob() needs to depend on internals
of the dirstream implementation.

Also, no implementation function shall set errno to 0, so all functions calling
readdir() need to save errno and restore it on success.