Disable cancellation during scandir().
Whether it is strictly conforming to expose this fact to the callbacks I don't
rightly care. The point of this is to obviate the need to handle cancellation
cleanup, and if the callbacks enable cancellation themselves, then they have to
know what they are doing.

More important than strict conformance is safety of the implementation, and
having an implementation that sometimes leaks memory even if the application
does everything right is not good. It is acceptable if the application does
something dumb like re-enabling cancellation in the callbacks. OK, maybe later I
add correct cancel handling here.