Simplify twalk().
Anderson trees have the property that all singleton nodes have a right child.
Therefore any node that doesn't have a right child must be a leaf. It is not
necessary to explicitly check if there is no left child, because as long as the
tree is constructed correctly, no left child can exist if no right child does.