Better implement strchr() and strrchr() semantics.
I read the spec such that these functions should convert the argument to type
char. This means it is possible for the number to be negative, if char is
signed. So I now ensure the argument is in the correct value range for the
portable code, and limit the value properly to 1 byte in the extension code.