Add simple parts of c11 threads.
All of the implemented functions wrap existing pthread functions. So the main
part of the task was to rename the functions in a namespace-safe way, so it can
be called from ISO-C code. And some of these actually *are* just the ISO-C
functions, so the implementation is just a weak alias.