Make endian macros require arith arg.
In the case where the macros don't do anything, they were just type casts
before. But casts can do more than implicit conversion can. In particular, casts
can convert pointers to integers. So I need to add an operation to the macros
that can only be performed if the argument is of such a type that an implicit
conversion to the destination type would be possible.

That conversion is possible for all arithmetic types. A no-op that is possible
with all arithmetic types is adding 0. But that is not necessarily a no-op with
floating-point types, and it is also possible with pointers. But multiplying 1
is always a no-op, so the likelyhood is high that this operation is removed
entirely. And it is impossible with pointers. So that's why I went for
multiplication with 1.