__libc_realloc: Call correct allocator.
When using __libc_realloc(), we must call __libc_malloc() in the background, so
we always get a chunk allocated by the correct allocator to use __libc_free()
on. This already happens further down when implementing the alloc+copy strategy,
so it also must happen at the top when p is NULL. Otherwise, malloc() might be
overloaded with something else and use a different allocation strategy.