printf: Ensure file is in write mode.
This was a circuitous solution to a different problem: When sprintf() is called
with an empty format string, it should just write a null byte and return.
However, mine didn't do that; it wrote nothing. vsnprintf() sets up a file that
will always null-terminate the string buffer if the write function is called,
but it wasn't called at all. Reason is that with an empty format string, not a
single byte is ever written to the file, and so the direction is never set, and
so the fflush() at the end of vfprintf() does nothing. With this, it will call
the write function at least once to flush the buffer, thus ensuring the
null-byte is written.

Semantically, calling printf() with an empty format string on a file not open
for writing is an error, and should be recorded as such. It is still a write
operation after all.