Promote killlock to all sigs blocked.
pthread_kill() blocks all signals before taking the killlock. This is necessary
to block cancellation in that function, and I don't have (and don't want to
have) a function to separately block cancellation and synccalls. pthread_kill()
must block cancellation, because otherwise if the thread is asynchronously
cancelled while pthread_kill() (called from pthread_cancel()) holds the
killlock, pthread_exit() (called from cancel_handler()) will deadlock.

OK, but that means pthread_kill() blocks SIGSYNCCALL while locking the killlock.
If another thread holds the killlock, but was interrupted by a synccall, then
this will deadlock. The only way to avoid this is to allow synccalls in
pthread_kill(), or else block all signals in the other three functions. I think
that is the smaller change.