Fix race condition in aio_suspend().
The previous design was like this:
    - __aio_notify contains the TID of the last thread registering for
      aio_suspend().
    - AIO worker sets __aio_notify to 0, and if there was something else there,
      performs a broadcast wake.

This has two problems:
1. Two threads trying to suspend at the same time can livelock each other,
   constantly overwriting the other's TID before it can start waiting.
2. An AIO operation finishing between the aio_error() check and actually writing
   the TID to __aio_notify is not noticed.

musl "solves" this problem by making aio_suspend() not overwrite their
equivalend of __aio_notify if it's already nonzero. However, that approach is
susceptible to an ABA problem instead. I think I shall have to make a bug
report.

Instead of these, I now replace __aio_notify with a serial number and waiters
bit. aio_suspend() crucially reads the serial number *before* checking on the
cbs. The AIO worker always increments the serial number and clears the waiters
bit when it finishes. An AIO worker finishing between aio_suspend() checking on
the cbs and setting the waiters bit is noticed by the a_cas() there now, and an
AIO worker finishing between aio_suspend() setting the waiters bit and actually
starting to wait is noticed by SYS_futex, because the waiters bit is cleared.

Since __aio_notify cannot take infinitely many states, this is technically still
susceptible to an ABA problem, but it would need a thread to not notice 2
billion updates to the variable for this to actually occur, which I will file
under "reasonably unlikely".