Add GNU hash support for __vdsosym().
I read on the musl mailing list that apparently on some architectures only the
GNU hashtab is emited. Not sure why, and there are certainly enough examples of
why that is a bad idea. But alright, I will add a function that counts the
symbols according to the GNU hash table.