Split futex_wait_cp() into its own file.
Changing futex_wait() into a cancel point was a small source change, but the
linkage changed significantly. futex_wait_cp() pulls in a host more things that
not every caller of futex_wait() wants. In the end, I decided the duplicated
code was small enough to be worth the optimization.