Fix binmap accesses.
Wow, this never worked, and I only noticed when I straced it. I never set the
necessary bit in binmap when binning a chunk, so the allocator always saw an
empty heap and went to the OS for more memory. And when unbinning a chunk, I
wrote the line to clear the bit wrong and instead cleared all the other bits.
That was not clever.