fnmatch: Test for exhaustion before reading pattern.
The loop to find the end of a sub-bracket may run into the end of the pattern
first. In that case, we get len >= plen when looking whether we found the end
bracket, which means we must not read the pattern there. This is of course only
a problem with broken bracket expressions. Why must POSIX be so complicated?