glob: Add support for using d_type.
Many filesystems on Linux now have support for passing d_type in the dirents.
Since all the allocated paths already have an extra byte at the end for
GLOB_MARK, I use that byte to communicate from process_name() to glob() whether
I already know a path exists. Unfortunately, that byte is also used to hold an
appended '/' if the pattern ends in '/', so in that case I synthesize the type.
I set it to UNKNOWN_EXISTANCE because I don't know whether this path actually
exists. Even if I know it exists without the '/'.

I'm using the same code as before for the UNKNOWN_EXISTANCE case, as well as for
KNOWN_EXISTS and GLOB_MARK. There is an unavoidable race condition when using
lstat(), since the FS might have changed between the readdir() and the lstat(),
but such is life.