Simplify *trylock().
In the simple case, that function doesn't need any retries beyond the ones in
a_cas() on the RISC architectures. The lock can only have three possible values,
namely 0, EBUSY, and WAITERS | EBUSY. So returning the return value of the first
a_cas() with the waiters bit masked off is always the correct value and the
correct behavior.