Add totalorder() and totalordermag().
I read the specification and at first thought a simple signed comparison on the
representation would be enough. But this is not quite true, since floats use
signed-magnitude representation whereas integers use two's complement
representation. Therefore, the order is opposite in the negative realm. But this
only matters if both inputs are negative, otherwise the comparison is already
correct.

Therefore the logic is to turn the representation into a signed integer of the
same length and perform a normal comparison, flipping the result if both inputs
are negative. This is a bit challenging for long double, but I think I did it
correctly. For long integers, the best view is to treat only the leading word as
signed and the rest as unsigned.

For totalordermag(), I only have to ignore the sign bit, which I do in the
simple cases by shifting it out. This also removes the possibility for either
input to be negative, so I never need to flip the result.

For the long double versions, if LDBL_MANT_DIG==53, this code technically
commits an aliasing violation with the pointer cast, but I have no other choice.
I cannot read the values into my own variables since the inputs might be SNaNs
and I must not raise an invalid exception in these functions.