Remove "/" failure case from glob().
When the glob pattern is just "/", then at the very end, a duplicate of an
anonymous root string is allocated. This is done just so that globfree() can
free all strings. But it presents an unnecessary failure case. So I now change
the root string to a named constant. In order for globfree() to be able to
recognize it, I also need to move it to the same function. But there was never a
good reason to have globfree() in a different file, since glob() by itself
already pulls in free().