Fix sign issues with timezone offset.
The sensible format for the timezone is "seconds east of DST", because that is
the amount of time that has to be *added* to UTC to get local time. But POSIX's
americacentrism caused them to flip the signs. Now the offset from the zoneinfo
files is actually the right sign, but the value I have to have in __timezone is
the wrong sign. I now fix this by flipping the sign when getting the POSIX
timezones, and then adding the offset in localtime_r().