Fix infinite binary search for zoneinfo.
The midpoint of end and start is of course the average of the two, or their sum
halved, not their difference halved. And the loop has to terminate when there is
one element remaining, not zero, because we will always keep one element around,
which is the element we wanted to find in the first place.