ldso: Correct depth-first search algorithm.
The algorithm I had written would work for a dependency tree. Unfortunately,
shared library dependencies often have multiple nodes depending on the same
library, and so multiple nodes link to the same node, and so our dependency
graph is not a tree.

One concrete example: main app depends on both libX11 and libc, libX11 depends
on libc. If the dependency list on the main app lists the libc first, then the
previous algorithm would push main app, libc, and libX11 to the stack, then
consider libX11, but *not* push libc since it is already marked. And then would
initialize libX11 *before* libc. In this case, since libc doesn't have
initializers or finalizers, it doesn't matter, but it is still wrong.

One way to fix it would have been to *not* mark libraries that are being pushed,
but only mark them when they are seen at the top of the stack. This would allow
us to push libc when considering libX11 in the above example. But the problem is
that this would allow pushing the same library multiple times (actually
unbounded), so now I cannot give an upper bound for the stack size.

So I am instead adopting musl's algorithm. That algo is to successively push
each unmarked dependency. So we first consider the main app, push its first
dependency, then consider that dependency and push *its* first dependency, and
so on, until we arrive at a leaf node. At which point we add the leaf node to
the queue, then push the parent's *second* dependency. A non-leaf node is only
added to the queue once all of its dependencies have been processed.

However, unlike musl, I do not allocate the counter separately. Instead, I widen
the "mark" field to a size_t and use it to count the position in the dependency
list.

In the above example, this would lead to us either first pushing and considering
libc, which is a leaf node, or to us pushing libX11, and then libc because it is
libX11's dependency. In either case, no matter how the dependencies are ordered,
libc is added to the queue before libX11.