Use call/pop for i386 sysinfo.s.
clang uses call next + pop to get the current IP, and this way uses one less
instruction. And according to the optimization guide, the return address stack
implementation understands that call next is not a subroutine call, and has done
for some time. So I am returning to using the simpler code here.