i386 sysinfo: Put oldsyscall and __get_sysinfo() in different sections.
I am compiling with -ffunction-sections to tell the compiler to do the same, but
fail to do so in my own code. Here it is probably OK to tell the linker that
these three bytes can be put elsewhere if need be. Might actually be an
optimization for large static links.