Handle O_APPEND and O_CLOEXEC in fdopen() and fopen().
I was already calling __sys_open() with those flags if requested. However,
__openat() has additional code to call fcntl() again, because some old kernel
versions ignore O_CLOEXEC altogether. This was not handled in this code path so
far.

POSIX also says that fdopen() with an append mode should set O_APPEND on the FD.
Now, this isn't a requirement, but some application may still want to see this
behavior, and I may document it in future (once I actually add documentation).
Setting O_APPEND in Linux can only be done by calling fcntl(F_SETFL), but unlike
with CLOEXEC, there is more than one possible changable flag that may be set on
the file descriptor. In particular something like O_NONBLOCK. So I have to query
the FD flags. Which may fail and provide me with a reason to leave the function
early.

Doing this in __fdopen_flg() would be overkill, since for fopen() I already know
the FD was opened correctly. So I am adding this code to fdopen() only.