Deduplicate printf code from __fmt_fp().
I had noticed that while fmt_d() and fmt_x() have to be capable of dealing with
uintmax_t in printf(), they don't have to do so in __fmt_fp(), where they will
only ever be called with arguments of type int or uint32_t.

At the moment, int is 32 bits on all supported platforms, and this is unlikely
to change. Even if it changes, the value given is a binary or decimal exponent,
and the value range for these fits in 32 bits after taking the absolute value.
We would need a future floating-point type with a 32-bit exponent, which I
highly doubt we will ever get. Even IEEE quad precision uses a 15 bit exponent,
which already enables a range that is far beyond reasonable.