__fmt_fp(): Avoid bad block of zeroes.
If rounding propagated through the entire number in fmt_decfloat(), then
rounding would end with r == a + 1 and z == r + 1 == a + 2. And *r == 0. So the
number as presented in the end would have one block of zeroes in it. This could
possibly mess up exponent calculations later. It's not supposed to have that.