Properly document PPC64 baseline kernel as v3.13.
For PowerPC64, I do not (and will not) support the ELFv1 ABI. That ABI is dumb.
The ELFv2 ABI is a lot better. But the change is not invisible to the kernel.
The kernel must support the ABI for the entry point and the signal functions to
work. So ELFv2 programs cannot work at all on older kernels.

As far as I can tell, ELFv2 support was added into Linux v3.13. This means, the
baseline Linux version for PPC64 is v3.13.

I should probably add a document listing this all per architecture.