Add getrlimit() and setrlimit().
Aarch64 dances to its own tune, again. It doesn't have either of these as system
calls, it only has prlimit64. So the fallbacks to the older syscalls need to be
conditional on them existing.

As for the older syscalls, from Linux 2.6.0 to 2.6.10, they only acted on the
current thread. Since prlimit64 was added later than the move of the rlimits to
the process level, a successful call to that syscall means the limits were
adjusted on the process level.

Now, for getrlimits() there is no issue, since all threads are supposed to have
the same rlimits. But for setrlimits(), this means we have to do a synccall
whenever we have to use the setrlimits syscall.

Since also the older syscalls may use smaller types for the limits, prlimits64
is to be preferred.