Define maths extension macro for totalorder().
Compiling for PPC32 gave me an error, indicating that totalorder() was not
declared when compiling totalorderl(). That is because math.h only declares
those functions in the presence of a feature test macro that was not defined.
Since I want to cross-check whether implementation and definition use compatible
definitions, I have to define that macro for all the implementations.