Add getrusage().
Boring stuff. But I needed to take care to initialize the two timevals
correctly. Yes, on 64-bit architectures this is completely superfluous, on
32-bit architectures it is actually a necessary sign extension.

Using 32-bit timevals for rusage is not expected to be a major issue, since very
few programs will have runtimes on the order of 68 years at 100% CPU.