Simplify x86_64's __tlsdesc_dynamic.
By using some parts of memory directly, I don't need to use RCX, so spilling and
restoring that goes away. And a bit more thought managed to swap the uses of RDX
and RAX, which meant in the end I could use an instruction sequence that was one
instruction shorter.