Optimize glob(): Copy pattern once.
For the case where we need to match a pattern on all files in the directory, we
need to copy the pattern to its own buffer, so we can push its null-terminated
version into fnmatch(). And the previous version of that code did that for every
file found in every candidate directory, which is just unnecessary.