Fix glob() heap corruption.
This was dumb. After assuring myself that result == &noalloc is only possible in
the case where only the root was matched, I had completely neglected to look at
the if-selection directly above that place, which also uses noalloc for the
GLOB_NOCHECK option to just return the pattern. So for that I need to allocate a
single pointer and a copy of the pattern.

Of these two, only the pointer is technically unnecessary, if I somehow invented
a way to have a non-allocated pointer list with allocated pointers inside. But
that would make the memory management even harder than before, and the heap can
probably take those extra 4 or 8 bytes. Especially in that place, where almost
surely other allocations have taken place and been freed before.

As it was, the code would add the pointer "pat" to the output list of strings
and globfree() would then free it, but that is of course improper, since pat is
not owned by the glob library and not necessarily allocated in the first place.
And indeed, the actual string could be at the end of its life directly after
returning.