Make chmod() go through fchmodat() and add fallbacks.
All other system call wrappers are implemented by having the older functions
call the newer functions, then implementing the newer functions in terms of
calling the older system calls when conditions allow. This does duplicate some
work, but the fallback code makes the newer functions more usable on older
kernels, so it must be present anyway, and then the older functions may as well
use them.

fchmod() stays the way it is, because SYS_fchmod is currently defined by all
architectures. If that changes, it too will have to fall back.