Add PPC-specific clock_gettime().
PPC and PPC64 have a VDSO that follows its own weird calling convention,
where success is indicated by CR0.SO. I thought about several ways to
integrate this, but ended up at no sensible way to do it without moving
everything to assembler.

First idea was an arch-specific wrapper around the indirect call. But
that would only work if the syscall function didn't already take care of
CR0.SO. So I would also have to use a special syscall fallback in a way
that's not used anywhere else.

Second idea was to just wrap the VDSO call, but that would make every
clock_gettime call a double-indirect call. Just for the bloody error
code!

Finally, I thought of just always taking the absolute value of the
result and treating it as an error if nonzero, since clock_gettime()
only returns 0 on success. Possible but hacky.

The best option left really is an arch-specific implementation of the
idea. And since I cannot guarantee nobody is mucking with anything in C,
it needs to be assembler.