Add release barriers to clock_gettime().
After a store instruction, the store is added to a store queue, to be
flushed at some point in future. But for other cores to see the value it
must be at least in the L1 cache. That will still happen without the
sync, but I have no control over the timing.

Ultimately, it doesn't really matter, since another core reading an
outdated value will calculate the same result in cgt_dispatch(). But it
is semantically more correct this way.