Improve AIO cancellation.
Oh boy, this took some doing.

The AIO worker has to wait for preceding writes to finish if its operation is a
sync or a write to an append FD. For that wait, it uses pthread_cond_wait().
That function releases the queue lock and waits for the condition to be
signaled, but will *always* reacquire the lock before leaving.

aio_cancel() would cancel an AIO worker, and then, hogging the queue lock, wait
for the running futex to change its value 0.

Until this commit, if a worker was cancelled waiting for preceding writes to
finish, the AIO cleanup function was not run at all, and aio_cancel() would
starve. However, simply pushing the cleanup function would not work: Since
pthread_cleanup_push() has to reacquire the mutex that aio_cancel() is hogging,
this would deadlock.

This isn't a problem anyone other than aio_cancel() can remedy: It has to
release the queue lock while waiting for the task to finish.

I already see a new problem to fix, though: This means that aio_cancel() loosens
its grip on the task list, and therefore the task list can change while
aio_cancel() is iterating. And that will not end well for anyone.