Add fmtmsg.
I surveyed existing implementations for guidance on two topics:

1. What do if class==MM_NULLMC?
2. What do with class at all, besides selecting the destination?

Surveyed implementations: glibc, musl, dietlibc, newlib, bionic.

The latter three don't have this interface at all. That's a bit of an API
failure, isn't it? It was added to POSIX in 1994 and became a base requirement
in 1997, and still I couldn't find a single user of it in all of Debian's
codebase, and those three projects didn't think to add it at all, because none
of their users have it either.

glibc and musl both don't use class at all, except for selecting the
destination. This also means they both return MM_OK without doing anything when
class == MM_NULLMC. Both construct the message with a function of the printf() family.
But glibc selects syslog instead of /dev/console when MM_CONSOLE is set.

I decided to forego the printf() path, to save codesize if printf() is not
otherwise referenced. The generated message is created by pasting strings
together, which I can just do with writev().

Since the only other implementations I could find also retry writing on error, I
do the same.