Improve AIO cancellation further.
With aio_cancel() letting go of the queue lock, it cannot keep a reference to
the task around after starting to wait, because the task's cleanup function can
win the race to reacquire the lock, and will only let go of it again when the
task is unlinked from the list and on track to end (thus ending the lifetime of
the task struct).

So aio_cancel() has to do its business without looking at the task again, and at
the moment, I do this by looking at the aiocb. But this presents a race
condition, because technically the application is free to free or reuse the
aiocb as soon as its error code has been changed from EINPROGRESS.

I think I will have to add yet more synchronization here, so in the cancellation
case, the worker doesn't end before all cancellers (there can be multiple ones)
have seen the task's return value. Not quite sure what the right primitive here
is. Barriers? Then I'd have to actually implement them, finally.