Add wordexp().
Idea from musl: Use the shell to perform the actual evaluation. This does not
permit me to implement WRDE_UNDEF, since the expansion takes place in another
process, but otherwise it is a pretty slick solution.

The idea is to execute the command

sh -c 'eval printf "%s\\\\0" x $1 $2' $expr $redir

where redir is conditionally the empty string or the string "2>/dev/null" and
expr is the given expression. The shell will unescape the format string once and
replace $1 and $2 with the arguments, and perform word splitting, leading to the
command

eval printf "%s\\0" x <split expr> <redir>

And then eval will will evaluate the entire thing again, unescaping the format
string once more, evaluating the already split expression, and honoring the
redirect command. After the eval, what is left is

printf "%s\0" x <split expr results>

and that is then executed. The definition for printf says that in this case, all
the words are output, terminated with null bytes.

The function of the leading x is to permit me to tell an empty split result from
a split result containing only an empty string. One will cause the output of
"x\0", and the other will cause "x\0\0".

After this, I only read out the bytes from the pipe, appending each to the
current word and breaking at null bytes, appending pointers to each word to a
vector. The rest is a little bit of management.

What's still left to be done is the implementation of scan_expr() to reject bad
characters as per spec. But that is for the future.