Add statvfs() and fstatvfs().
I don't like the duplicated code, but I honestly don't know how to abstract it
better, and it is only one line anyway.

I decided against using the spare words at the end of the kernel's struct
statfs64 to store the additional member. If the kernel ever does use those
fields, my own field won't be in the way.