tsearch(): Return already-inserted node.
If a key is found in the tree, tsearch() is supposed to return that node. For
that reason, if cmp() returns 0, it should also set the return pointer.

Of course, with that we no longer have a signal whether the lowest level created
a new node, so I can save a branch.