x32/syscall_arch.h: Prevent undefined behavior.
Technically, pointer arithmetic on void pointers is undefined, even if the
arithmetic in question is just adding zero. But the type switch I want to make
here can be made more easily with a ternary expression: Since 0ULL is also a
null pointer constant, a ternary expression involving it and a pointer to
anything will have the other pointer as result type, whereas if the other
expression is an integer, the "usual arithmetic conversions" apply, and the
result will be of type unsigned long long (since no integer type with higher
conversion rank exists).