Replace masks with shifts.
The idea of getting rid of the sign bit with shifts rather than with explicit
masks is done in other places in this file, and is less text to read, and easier
to get right. While the constants written were correct, it would be so easy to
have an f to many or too few in that long string. And it's all the same to the
compiler, anyway.