Fix const-correctness macros for C23.
A recent thread on the musl mailing list made me look at my own implementation
of this. And I had to fix a few things:

The selecting expression "1?(s):(void *)0" unfortunately has a null pointer as
third operand, and therefore its result type is the type of s, not void* or
const void*, as I'd hoped. To get that result, I have to use a non-null void
pointer as third expression. The simplest non-null pointer I can procure at
short notice is the empty string. So the const-query expression must be

1 ? (s) : (void *)""

For the mem* functions, that expression already has the type I want to cast the
return value to. But for the str* functions, I must also convert that to
character pointer type. I saw no way to do that but with a generic selection,
and so that is what I implemented.