Fix buffer overflow in __donate_malloc_memory().
I guess I wanted to either write "size_t" or "char *" there, and the two ideas
combined into the unhealthy "size_t *". In this case, it causes malloc() to take
control of 4 or 8 times as much memory as was actually given to it. In this
case, char * is the better type.

Man, if I ever try porting to something that has colored memory, stuff like this
will be my undoing.