__fmt_fp: Fix off-by-one errors in post-radix digit calculation.
I am starting to wonder if just increasing rp by 1 is a good idea. At the
moment, z is a pointer to 1 past the end of the number, but rp is a pointer to 1
in front of the post-radix digits. This means the number of digits in front of
the radix point is given by countdig(*a) + (rp - a) * 9, and the number of
digits after is given by (rp - z - 1) * 9. So the previous expressions were
wrong for the current model.