Add preliminary ARM support.
I wanted to see how much work it is to add an architecture to the library, and
it turns out not a lot, but there is some stuff to do.

First order of business was to create all the files under arch/arm/bits. Much of
this is kernel interface. The fenv stuff comes down to architecture again. Next
task was to create the stuff under arch/arm. Much of which is also kernel
interface.

The crt1s.s and rcrt1s.s also have to be created, as I require them for all
builds at the moment. As is the fenv stuff, which thankfully was an easy
modification of the aarch64 stuff.

The string functions are a necessity. In this case, the ABI requires the
existance of further functions, and the ISO-C functions can be implemented in
terms of them. So I wrote them accordingly.

The big problem is to initialize the atomic functions. It turns out ARM32
doesn't have any atomic functions in the baseline ISA (ARMv4). Instead, the
kernel provides helper functions with the necessary semantics for the platform.

As far as I could ascertain, the following are the important differences between
the architecture versions for my purposes (there are more differences listed,
but they don't matter to me):

ARMv4: Baseline
ARMv5: Added BLX (so prefer to call external functions with BLX if ARMv4-compat
        is not required)
ARMv6: Added LDREX and STREX, and standardized CP15. Therefore, in ARMv6 and
        above, CAS can be implemented without the use of the kuser helpers.
ARMv6T2: Added LDREX and STREX to Thumb
ARMv7: Added DMB ISH operation to ARM and Thumb.

For now, everything is using runtime-selected code. The actual lock operations
call function pointers initialized in __set_thread_area(), and those go to the
kernel on v4 and v5, or to our own definitions on v6 and v7.

However, if compiling for ARMv7 or greater, the runtime code selection can be
entirely eliminated, and the correct instructions can be inlined.

I have written all the source files in UAL, which should allow them to compile
as ARM and Thumb, but I haven't yet had luck making the compiler create Thumb
code. We will see.

The runtime code selection means that ARM's __set_thread_area() needs to know
some aux vectors. I did not want to add them as useless arguments to the
function, so I added support for getauxval() from glibc, which allows
__set_thread_area() to query hwcap and platform. But this adds a dependency that
wasn't there before, so now the initialization of __auxv has to happen before
__set_thread_area() can be called. And that means I had to shuffle some code
around in the dynamic linker especially (where decoding the stack used to only
happen in linker_main.c, but now some of it happens in bootstrap.c).

As for architecture variants, I only support VFP for hardware floating point
(beyond compiler support). Switching this at runtime would be unreasonable,
though.